+++ packages/@expo/metro-config/src/serializer/exportHermes.ts
@@ -96,8 +96,13 @@ async function directlyBuildHermesBundleAsync({
     if (minify) {
       args.push('-O');
     }
+
     if (map) {
       args.push('-output-source-map');
+    } else {
+      // When source maps are not enabled, Hermes will generate maximum debug information.
+      // This causes the bundle to inflate in size. We can disable debug information to reduce the size.
+      args.push('-g0');
     }
 
     debug(`Running hermesc: ${hermesCommand} ${args.join(' ')}`);